dependency enhancements & cleanup (#4634)
* replace async-iterable-queue by it-pushable

* bump typescript

* housekeeping

* some tree-shaking

+++ packages/core/src/messages/acknowledgementChallenge.ts
@@ -63,7 +63,11 @@ export class AcknowledgementChallenge {
 }
 
 function verifyChallenge(pubKey: PeerId, signature: Uint8Array, challenge: HalfKeyChallenge): boolean {
-  return secp256k1.ecdsaVerify(signature, hashChallenge(challenge), unmarshalPublicKey(pubKey.publicKey).marshal())
+  return secp256k1.ecdsaVerify(
+    signature,
+    hashChallenge(challenge),
+    unmarshalPublicKey(pubKey.publicKey as Uint8Array).marshal()
+  )
 }
 
 function hashChallenge(ackChallenge: HalfKeyChallenge): Uint8Array {
